patch: extract FbmSettings to its own module
